Service: add atomic ratings. Ratings are not integers.

+++ src/main/java/pl/arciemowicz/rankingbackend/controller/RankingController.java
@@ -33,7 +33,7 @@ public class RankingController {
     }
 
     @RequestMapping(value = "rate/movie/{movieId}", method = RequestMethod.PUT)
-    public Rate addMovieRating(@PathVariable("movieId") String movieId, @RequestBody Double rate) {
+    public Rate addMovieRating(@PathVariable("movieId") String movieId, @RequestBody Integer rate) {
         return ratesService.addRating(Type.MOVIE, movieId, rate);
     }
